Policy Install and Test

+++ app/Http/Controllers/TaskController.php
@@ -72,11 +72,8 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
-        //
-
+        $this->authorize('delete', $task);
         $task->delete();
-
-
         return response()->json(null, 204);
     }
 }
